requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
firebase-admin==6.5.0
google-cloud-language==2.13.4
google-api-python-client==2.149.0
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.3.1
requests==2.32.3

fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
firebase-admin==6.5.0
google-cloud-language==2.13.4
google-api-python-client==2.149.0
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0
lxml_html_clean==0.3.1
requests==2.32.3
